+++ Professions/Commands/AddCommand.cs
@@ -5,18 +5,20 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using DaLion.Professions.Framework;
 using DaLion.Shared.Commands;
 using DaLion.Shared.Extensions;
 using DaLion.Shared.Extensions.Collections;
 using DaLion.Shared.Extensions.SMAPI;
+using StardewValley.Constants;
 using StardewValley.Menus;
 
 #endregion using directives
 
-/// <summary>Initializes a new instance of the <see cref="AddProfessionsCommand"/> class.</summary>
+/// <summary>Initializes a new instance of the <see cref="AddCommand"/> class.</summary>
 /// <param name="handler">The <see cref="CommandHandler"/> instance that handles this command.</param>
 [UsedImplicitly]
-internal sealed class AddProfessionsCommand(CommandHandler handler)
+internal sealed class AddCommand(CommandHandler handler)
     : ConsoleCommand(handler)
 {
     /// <inheritdoc />
@@ -24,7 +26,7 @@ internal sealed class AddProfessionsCommand(CommandHandler handler)
 
     /// <inheritdoc />
     public override string Documentation =>
-        "Add the specified professions to the player. Does not affect skill levels.";
+        "Add the specified professions to the player without affecting skill levels. Can also be used to add masteries to the specified skills using the keyword \"mastery\".";
 
     /// <inheritdoc />
     public override bool CallbackImpl(string trigger, string[] args)
@@ -33,6 +35,49 @@ internal sealed class AddProfessionsCommand(CommandHandler handler)
         {
             this.Handler.Log.W("You must specify at least one profession.");
             return false;
+        }
+
+        if (args[0].ToLower() is "mastery" or "masteries")
+        {
+            args = args.Skip(1).ToArray();
+            foreach (var arg in args)
+            {
+                if (string.Equals(arg, "all", StringComparison.InvariantCultureIgnoreCase))
+                {
+                    foreach (var skill1 in Skill.List)
+                    {
+                        if (skill1.CanGainPrestigeLevels())
+                        {
+                            continue;
+                        }
+
+                        Game1.player.stats.Set(StatKeys.Mastery(skill1), 1);
+                        this.Handler.Log.I($"Mastered the {skill1} skill.");
+                    }
+
+                    Game1.player.stats.Set(StatKeys.MasteryExp, MasteryTrackerMenu.getMasteryExpNeededForLevel(5));
+                    return true;
+                }
+
+                if (Skill.TryFromName(arg, true, out var skill2))
+                {
+                    if (skill2.CanGainPrestigeLevels())
+                    {
+                        this.Handler.Log.I($"{skill2} skill is already mastered.");
+                        return true;
+                    }
+
+                    Game1.player.stats.Set(StatKeys.Mastery(skill2), 1);
+                    Game1.player.stats.Set(
+                        StatKeys.MasteryExp,
+                        MasteryTrackerMenu.getMasteryExpNeededForLevel(MasteryTrackerMenu.getCurrentMasteryLevel() + 1));
+                    this.Handler.Log.I($"Mastered the {skill2} skill.");
+                }
+                else
+                {
+                    this.Handler.Log.I($"Ignoring unknown vanilla skill \"{skill2}\".");
+                }
+            }
         }
 
         var prestigeArgs = args.Where(a => a.ToLower() is "-p" or "--prestiged").ToArray();
@@ -131,17 +176,23 @@ internal sealed class AddProfessionsCommand(CommandHandler handler)
     /// <inheritdoc />
     protected override string GetUsage()
     {
-        var result =
+        var sb =
             new StringBuilder(
                 $"\n\nUsage: {this.Handler.EntryCommand} {this.Triggers[0]} [--prestige / -p] <profession1> <profession2> ... <professionN>");
-        result.Append("\n\nParameters:");
-        result.Append("\n\t- <profession>\t- a valid profession name, or `all`");
-        result.Append("\n\nOptional flags:");
-        result.Append(
+        sb.Append("\n\nParameters:");
+        sb.Append("\n\t- <profession>\t- a valid profession name, or `all`");
+        sb.Append("\n\nOptional flags:");
+        sb.Append(
             "\n\t-prestige, -p\t- add the prestiged versions of the specified professions (base versions will be added automatically if needed)");
-        result.Append("\n\nExamples:");
-        result.Append($"\n\t- {this.Handler.EntryCommand} {this.Triggers[0]} artisan brute");
-        result.Append($"\n\t- {this.Handler.EntryCommand} {this.Triggers[0]} -p all");
-        return result.ToString();
+        sb.Append("\n\nExamples:");
+        sb.Append($"\n\t- {this.Handler.EntryCommand} {this.Triggers[0]} artisan brute");
+        sb.Append($"\n\t- {this.Handler.EntryCommand} {this.Triggers[0]} -p all");
+
+        sb.Append($"\n\nAlternative usage: {this.Handler.EntryCommand} {this.Triggers[0]} mastery <skill1> <skill2> ... <skillN>");
+        sb.Append("\n\nParameters:");
+        sb.Append("\n\t- <skill>\t- a valid vanilla skill name, or `all`");
+        sb.Append("\n\nExamples:");
+        sb.Append($"\n\t- {this.Handler.EntryCommand} {this.Triggers[0]} mastery farming");
+        return sb.ToString();
     }
 }